// Java/Tamin-Projects/eservice-lite/ui/src/app/modules/fac/models/alo-req-det-workshop-address.ts
import {City} from './baseinfo/city';
import {Province} from './baseinfo/province';
import {AloRequest} from './alo-request';
import {Branch} from './branch';
import {AloLocation} from './alo-location';
import {AloRefine} from './alorefine';

export class AloReqDetWorkshopAddress {
  detWorkshopAddressId: number;
  workshopId: string;
  workshopAddressDesc: string;
   cityCode1: string;
  workshopAddressPhone1: string;
  cityCode2: string;
  workshopAddressPhone2: string;
  cityCode3: string;
  cityCodefax: string;
  workshopAddressPhone3: string;
  refineName: string;
  refineLastName: string;
  refineNationalId: string;
  refineBirthDate: Date;
  branchParentCode: string;
  isOffice: string;
  requests: AloRequest;
  branch: Branch;
  name: string;
  email: string;
  faxNo: string;
  codePosti: string;
  aloRefineId: AloRefine;
  docLocationDesc: string;
  city: City;
  mobail: string;
  province: Province;
  nationalCode: string;
  locationId: AloLocation;
  docLocationAddress: string;
  legalResidence: string;
}